requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.9.2
Flask-PyMongo==2.3.0
flower==2.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
kombu==5.4.2
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.58.1
ordered-set==4.1.0
packaging==24.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
youtube-transcript-api==0.6.3
